CMake cleanup
CMAKE_CURRENT_SOURCE_DIR is not necessary in the dedicated target attribute
functions of cmake. Cmake will suffix any path token with the current
source path.

find_program is a useful utility to pick up binaries.

When SDL is installed the include path directives will not work - to mitigate
this problem on linux system the find_library(sdl2) is used. It lacks the
REQUIRED token - so users can still set a manually built a non-system SDL